# docs/swagger.yaml
basePath: /api/v1
definitions:
  errno.errno:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  model.Actor:
    properties:
      alias_name:
        type: string
      cover:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    required:
    - id
    - name
    type: object
  model.Movie:
    properties:
      code:
        type: string
      cover:
        type: string
      created_at:
        type: string
      director:
        type: string
      id:
        type: integer
      produce_company:
        type: string
      publish_company:
        type: string
      publish_date:
        type: string
      series:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - code
    - title
    type: object
  model.MovieActor:
    properties:
      actor_id:
        type: integer
      actor_name:
        type: string
      id:
        type: integer
      movie_id:
        type: integer
    type: object
  model.MovieInfo:
    properties:
      actors:
        items:
          $ref: '#/definitions/model.MovieActor'
        type: array
      movie:
        $ref: '#/definitions/model.Movie'
      tags:
        items:
          $ref: '#/definitions/model.MovieTag'
        type: array
    type: object
  model.MovieTag:
    properties:
      id:
        type: integer
      movie_id:
        type: integer
      tag_id:
        type: integer
      tag_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server celler server.
  license:
    name: MIT
    url: https://opensource.org/license/mit
  title: Garage API
  version: "1"
paths:
  /actor:
    delete:
      description: Delete actor
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete actors
      tags:
      - actor
    get:
      description: Get actors List
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errno.errno'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Actor'
                  type: array
              type: object
      summary: Get actors
      tags:
      - actor
    post:
      description: Create actor
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Create actors
      tags:
      - actor
    put:
      consumes:
      - application/json
      description: Update a actor by id
      parameters:
      - description: Actor object
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/model.Actor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/errno.errno'
      summary: Update a actor by id
      tags:
      - actor
  /actor/search:
    get:
      description: Search actors List
      parameters:
      - description: name search by name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errno.errno'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Actor'
                  type: array
              type: object
      summary: Search actors
      tags:
      - actor
...
  /movie:
    delete:
      consumes:
      - application/json
      description: Delete movies
      parameters:
      - description: default
        in: body
        name: default
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errno.errno'
            - properties:
                data:
                  type: object
              type: object
      summary: Delete movies
      tags:
      - movie
    get:
      description: Get movies
      parameters:
      - default: 50
        description: int valid
        in: query
        name: page_size
        type: integer
      - default: 1
        description: int valid
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errno.errno'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Movie'
                  type: array
              type: object
      summary: Get movies
      tags:
      - movie
    post:
      description: Create movies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/errno.errno'
      summary: Create movies
      tags:
      - movie
    put:
      consumes:
      - application/json
      description: Update a movie by id
      parameters:
      - description: movie object
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/model.Movie'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/errno.errno'
      summary: Update a movie by id
      tags:
      - movie
  /movie/info/:code:
    get:
      description: Get movies
      parameters:
      - description: movie code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errno.errno'
            - properties:
                data:
                  $ref: '#/definitions/model.MovieInfo'
              type: object
      summary: Get movies
      tags:
      - movie
  /movie/search:
    get:
      description: Search movies by code
      parameters:
      - description: movie code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errno.errno'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Movie'
                  type: array
              type: object
      summary: Search movies
      tags:
      - movie
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"